+++ src/module.ts
@@ -6,6 +6,9 @@ interface Options {
   prefix?: String
   i18n?: Boolean
   expires?: Number
+  loginUrl?: String
+  registerUrl?: String
+  userUrl?: String
 }
 
 const myModule: Module<Options> = function (moduleOptions) {
@@ -20,12 +23,11 @@ const myModule: Module<Options> = function (moduleOptions) {
   if (!options.guestPath) options.guestPath = '/'
   if (!options.prefix) options.prefix = 'authCustom_'
   if (!options.i18n) options.i18n = false
-  if (!options.expires) options.expires = 30
+  if (!options.expires) options.expires = 3600
 
   const pluginsToSync = [
     'plugins/index.ts',
     'plugins/storage.ts',
-    'middleware/index.js',
     'middleware/index.js',
     'util.ts',
   ]