+++ src/plugins/index.ts
@@ -1,37 +1,66 @@
 import { Plugin } from '@nuxt/types'
-import axios from 'axios'
 import { Storage } from './storage'
 
-// const options = JSON.parse(`<%= JSON.stringify(options) %>`)
-// interface loginData {
-//   acc?: String
-//   pass?: String
-// }
-
+const options = JSON.parse(`<%= JSON.stringify(options) %>`)
+const { registerUrl, loginUrl, userUrl, expires } = options
 const myPlugin: Plugin = (ctx, inject) => {
   const storage = new Storage(ctx)
   const allFn = {
-    async login(): Promise<Boolean | void> {
-      try {
-        const token = await axios.get('http://icanhazip.com/')
-        const userData = {
-          id: '0',
-          name: 'User',
-        }
-        storage.setUniversal('token', token.data)
-        storage.setUniversal('user', userData)
-        storage.setUniversal('isLogged', true)
+    register(data: any) {
+      return new Promise((resolve) => {
+        ctx.$axios
+          .$post(registerUrl, data)
+          .then((response) => resolve(response))
+      })
+    },
+    async login({ username, password }: any): Promise<Object | void> {
+      const res = await ctx.$axios
+        .$post(loginUrl, { username, password })
+        .then((response) => {
+          return response
+        })
+        .catch((_) => {
+          this.logout()
+        })
+      const { access_token, refresh_token, token_type } = res
+      if (access_token) {
+        const expired = new Date().getTime() + expires * 1000
+        storage.setUniversal('token', access_token)
+        storage.setUniversal('refreshToken', refresh_token)
+        storage.setUniversal('expiredAt', expired)
+        storage.setUniversal('tokenType', token_type)
 
-        return true
-      } catch (_) {
-        console.error('Log in ERROR!')
+        ctx.$axios.setToken(storage.state.token, storage.state.tokenType)
+        const userData = await ctx.$axios
+          .$get(userUrl)
+          .then((response) => {
+            return response
+          })
+          .catch((_) => {
+            this.logout()
+          })
+        if (userData) {
+          storage.setUniversal('user', userData)
+          storage.setUniversal('isLogged', true)
+          return { success: true }
+        } else {
+          this.logout()
+          return userData
+        }
+      } else {
+        this.logout()
+        return res
       }
     },
     logout(): Boolean | void {
       try {
         storage.removeUniversal('isLogged')
+        storage.removeUniversal('tokenType')
         storage.removeUniversal('token')
+        storage.removeUniversal('refreshToken')
+        storage.removeUniversal('expiredAt')
         storage.removeUniversal('user')
+
         return true
       } catch (_) {
         console.error('Log out ERROR!')
@@ -41,11 +70,7 @@ const myPlugin: Plugin = (ctx, inject) => {
       return storage.state.isLogged
     },
     getUserData(): unknown {
-      // try {
       return storage.state.user
-      // } catch (_) {
-      // console.error('Get user data FAILED!')
-      // }
     },
   }
   inject('authCustom', allFn)