+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### 1.0.3
 
-- add options `guestPath`
+- add option `guestPath`
 - remove vuex store replace with cookie
 - add function logout()
 - add return value and error log in api refrence
@@ -8,6 +8,9 @@
 - add userdata feature
 - add cookie expired option
 - add store state
+- add option `registerUrl`
+- add option `loginUrl`
+- add option `userUrl`
 
 ### 1.0.0